@@ -3,20 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "~4.2.4",
-    "@testing-library/react": "~9.5.0",
-    "@testing-library/user-event": "~7.2.1",
-    "json-server": "~0.16.2",
-    "mini.css": "~3.0.1",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/user-event": "^13.5.0",
+    "json-server": "^0.17.0",
+    "mini.css": "^3.0.1",
     "prop-types": "~15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-redux": "~7.2.1",
-    "react-router-dom": "~5.2.0",
-    "react-scripts": "3.4.4",
-    "redux": "~4.0.5",
-    "redux-devtools-extension": "~2.13.8",
-    "redux-thunk": "~2.3.0"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-redux": "^8.0.1",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
+    "redux-devtools-extension": "^2.13.9",
+    "redux-thunk": "^2.4.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
