@@ -9,7 +9,7 @@ import {
   BrowserRouter as Router,
   Route,
   NavLink,
-  Switch,
+  Routes,
 } from 'react-router-dom';
 import HomePage from './home/HomePage';
 
@@ -21,7 +21,7 @@ function App() {
           <span className="logo">
             <img src="/assets/logo-3.svg" alt="logo" width="49" height="99" />
           </span>
-          <NavLink to="/" exact className="button rounded">
+          <NavLink to="/" className="button rounded">
             <span className="icon-home"></span>
             Home
           </NavLink>
@@ -30,11 +30,11 @@ function App() {
           </NavLink>
         </header>
         <div className="container">
-          <Switch>
-            <Route path="/" exact component={HomePage} />
-            <Route path="/projects" exact component={ProjectsPage} />
-            <Route path="/projects/:id" component={ProjectPage} />
-          </Switch>
+          <Routes>
+            <Route path="/" element={<HomePage />} />
+            <Route path="/projects" element={<ProjectsPage />} />
+            <Route path="/projects/:id" element={<ProjectPage />} />
+          </Routes>
         </div>
       </Router>
     </Provider>
