@@ -1,12 +1,14 @@
 import React, { useEffect, useState } from 'react';
+import { useParams } from 'react-router-dom';
 import { projectAPI } from './projectAPI';
 import ProjectDetail from './ProjectDetail';
 
-function ProjectPage(props) {
+function ProjectPage() {
   const [loading, setLoading] = useState(false);
   const [project, setProject] = useState(null);
   const [error, setError] = useState(null);
-  const id = Number(props.match.params.id);
+  const params = useParams();
+  const id = Number(params.id);
 
   useEffect(() => {
     setLoading(true);
